Outline the fundamental structure of the application
1. index.html - The first page that the user is greeted with, is a page that serves as a greeting page.
Some random cool facts, or instruction manual on how to play CAH can be displayed here.
On clicking the sole button on the page, the user is redirected to the login page.
2. login.html -  In the login page, the user will be able to either join a random room, or create a new room. Whatever be it, once the user presses the button (assuming that the user has provided right credentials/has chosen to join a random room), the user is directed towards the next page.
3. room.html - In this page, the user has successfully joined the specified room, and is connected via the corresponding socket, to the other users present in the room.
This commit lays the rudimentary outline of the structure of the application from the user's perspective, and also serves as a demonstration of how web sockets seem to work well enough up until this point.