Modularize socket.io events
A new folder has been created in the src directory, for each socket
event. Previously, index.js was pretty long, thanks to all the socket
event handling. It gets real tiring, real fast to figure out what goes
where, and even more cumbersome, when it's not modularized enough.
This commit is a step towards modularizing the socket events. Soon
enough, each significant individual event will be housed in a file of
it's own, in an attempt to improve code readability, and make it easier
to scale, and debug.